+++ src/program/ui/LuaConsoleWindow.cpp
@@ -20,14 +20,48 @@
 #include <QtWidgets/QTableView>
 #include <QtWidgets/QPushButton>
 #include <QtWidgets/QDialogButtonBox>
-#include <QtWidgets/QVBoxLayout>
+#include <QtWidgets/QHBoxLayout>
+#include <QtWidgets/QFileDialog>
+#include <QtWidgets/QMenuBar>
+#include <QtWidgets/QMenu>
+#include <QtWidgets/QAction>
+#include <QtWidgets/QHeaderView>
+// #include <QtWidgets/QItemSelectionModel>
 
 #include "LuaConsoleWindow.h"
+#include "LuaConsoleModel.h"
 #include "../lua/Print.h"
+#include "../Context.h"
 
-LuaConsoleWindow::LuaConsoleWindow(QWidget *parent) : QDialog(parent)
+LuaConsoleWindow::LuaConsoleWindow(Context *c, QWidget *parent) : QMainWindow(parent), context(c)
 {
     setWindowTitle("Lua Console");
+
+    /* Main menu */
+    QMenu* scriptMenu = menuBar()->addMenu(tr("Script"));
+    
+    scriptMenu->addAction(tr("Add script file"), this, &LuaConsoleWindow::addScript);
+    scriptMenu->addAction(tr("Remove script file"), this, &LuaConsoleWindow::removeScript);
+    scriptMenu->addAction(tr("Clear all script files"), this, &LuaConsoleWindow::clearScripts);
+    scriptMenu->addAction(tr("Clear output window"), this, &LuaConsoleWindow::slotClear);
+
+    /* Table */
+    luaView = new QTableView(this);
+    luaView->setSelectionBehavior(QAbstractItemView::SelectRows);
+    luaView->setSelectionMode(QAbstractItemView::ExtendedSelection);
+    luaView->setShowGrid(false);
+    luaView->setAlternatingRowColors(true);
+    
+    luaView->verticalHeader()->hide();
+
+    luaModel = new LuaConsoleModel();
+    luaView->setModel(luaModel);
+
+    luaView->horizontalHeader()->setSectionResizeMode(0, QHeaderView::Fixed);
+    luaView->horizontalHeader()->resizeSection(0, 18);
+    luaView->horizontalHeader()->setSectionResizeMode(1, QHeaderView::Interactive);
+    luaView->horizontalHeader()->resizeSection(1, 80);
+    luaView->horizontalHeader()->setSectionResizeMode(2, QHeaderView::Stretch);
 
     /* Text Edit */
     consoleText = new QPlainTextEdit();
@@ -35,20 +69,14 @@ LuaConsoleWindow::LuaConsoleWindow(QWidget *parent) : QDialog(parent)
     consoleText->setMaximumBlockCount(10000);
     connect(&Lua::Print::get(), &Lua::Print::signalPrint, this, &LuaConsoleWindow::slotAppend);
 
-    /* Buttons */
-    QPushButton *clearButton = new QPushButton(tr("Clear"));
-    connect(clearButton, &QAbstractButton::clicked, this, &LuaConsoleWindow::slotClear);
+    QHBoxLayout *mainLayout = new QHBoxLayout;
 
-    QDialogButtonBox *buttonBox = new QDialogButtonBox();
-    buttonBox->addButton(clearButton, QDialogButtonBox::ActionRole);
-
-    /* Layout */
-    QVBoxLayout *mainLayout = new QVBoxLayout;
-
+    mainLayout->addWidget(luaView);
     mainLayout->addWidget(consoleText);
-    mainLayout->addWidget(buttonBox);
 
-    setLayout(mainLayout);
+    QWidget *centralWidget = new QWidget;
+    centralWidget->setLayout(mainLayout);
+    setCentralWidget(centralWidget);
 }
 
 void LuaConsoleWindow::slotAppend(QString qstr)
@@ -59,4 +87,29 @@ void LuaConsoleWindow::slotAppend(QString qstr)
 void LuaConsoleWindow::slotClear()
 {
     consoleText->clear();
+}
+
+void LuaConsoleWindow::addScript()
+{
+    QString filename = QFileDialog::getOpenFileName(this, tr("Lua script"), context->gamepath.c_str());
+    if (filename.isNull())
+        return;
+
+    luaModel->addFile(filename);
+}
+
+void LuaConsoleWindow::removeScript()
+{
+    /* Remove scripts from bottom to top */
+    QModelIndexList indexlist = luaView->selectionModel()->selectedRows();
+    std::sort(indexlist.begin(), indexlist.end());
+    
+    for (auto it = indexlist.crbegin(); it != indexlist.crend(); it++) {
+        luaModel->removeFile(it->row());
+    }
+}
+
+void LuaConsoleWindow::clearScripts()
+{
+    luaModel->clear();
 }